Move AtomicLazyCell into windows mod
